requirements.txt
pydantic
python-dotenv
langchain-openai
mem0ai
fastapi
uvicorn
langchain
crewai
openai-whisper
aiohttp
beautifulsoup4
SpeechRecognition
python-multipart
requests
# Existing requirements
